test(oauth): clean up client binding coverage
