+++ api/tests/oauth_code_client_binding_test.rs
@@ -5,105 +5,16 @@
 
 mod common;
 
-use axum::{
-    body::to_bytes,
-    extract::State,
-    http::StatusCode,
-    response::{IntoResponse, Response},
-};
+use axum::{body::to_bytes, extract::State, http::StatusCode, response::Response};
 use chrono::{Duration, Utc};
-use keycast_api::{
-    api::{
-        http::{oauth, routes::AuthState},
-        tenant::{Tenant, TenantExtractor},
-    },
-    bcrypt_queue::BcryptQueue,
-    handlers::http_rpc_handler::new_http_handler_cache,
-    state::KeycastState,
-};
-use keycast_core::{
-    encryption::{KeyManager, KeyManagerError},
-    secret_pool::SecretPool,
-};
-use moka::future::Cache;
 use nostr_sdk::Keys;
 use sqlx::PgPool;
-use std::sync::Arc;
-use tokio::task::JoinHandle;
 use uuid::Uuid;
-use zeroize::Zeroizing;
+
+use keycast_api::api::http::oauth;
 
 const ISSUED_CLIENT_ID: &str = "Code Binding Test Client";
 const REDIRECT_URI: &str = "https://test.example.com/callback";
-
-struct TestKeyManager;
-
-#[async_trait::async_trait]
-impl KeyManager for TestKeyManager {
-    async fn encrypt(&self, plaintext_bytes: &[u8]) -> Result<Vec<u8>, KeyManagerError> {
-        Ok(plaintext_bytes.to_vec())
-    }
-
-    async fn decrypt(
-        &self,
-        ciphertext_bytes: &[u8],
-    ) -> Result<Zeroizing<Vec<u8>>, KeyManagerError> {
-        Ok(Zeroizing::new(ciphertext_bytes.to_vec()))
-    }
-}
-
-async fn setup_pool() -> PgPool {
-    common::assert_test_database_url();
-    std::env::set_var("BUNKER_RELAYS", "wss://relay.test.example");
-    let database_url = std::env::var("DATABASE_URL")
-        .unwrap_or_else(|_| "postgres://postgres:password@localhost/keycast_test".to_string());
-    let pool = PgPool::connect(&database_url)
-        .await
-        .expect("Failed to connect to database");
-
-    sqlx::migrate!("../database/migrations")
-        .run(&pool)
-        .await
-        .expect("Failed to run migrations");
-
-    pool
-}
-
-fn create_test_auth_state(pool: PgPool) -> (AuthState, JoinHandle<()>) {
-    let bcrypt_queue = BcryptQueue::new();
-    let secret_pool = SecretPool::new(1);
-    let producer_handle = secret_pool.spawn_producer();
-    let tenant_cache = Cache::builder().max_capacity(10).build();
-    let key_manager: Arc<Box<dyn KeyManager>> = Arc::new(Box::new(TestKeyManager));
-
-    let auth_state = AuthState {
-        state: Arc::new(KeycastState {
-            db: pool,
-            key_manager,
-            signer_handlers: None,
-            http_handler_cache: new_http_handler_cache(),
-            server_keys: Keys::generate(),
-            tenant_cache,
-            bcrypt_sender: bcrypt_queue.sender(),
-            redis: None,
-            secret_pool: secret_pool.receiver(),
-        }),
-        auth_tx: None,
-    };
-
-    (auth_state, producer_handle)
-}
-
-fn test_tenant() -> TenantExtractor {
-    TenantExtractor(Arc::new(Tenant {
-        id: 1,
-        domain: "localhost".to_string(),
-        name: "Test Tenant".to_string(),
-        settings: None,
-        created_at: Utc::now(),
-        updated_at: Utc::now(),
-    }))
-}
 
 /// Insert a pending-registration authorization code issued to `ISSUED_CLIENT_ID`.
 async fn insert_authorization_code(pool: &PgPool, code: &str, user_pubkey: &str, email: &str) {
@@ -129,10 +40,10 @@ async fn insert_authorization_code(pool: &PgPool, code: &str, user_pubkey: &str,
 
 /// Redeem `code` at the token endpoint while presenting `presented_client_id`.
 async fn redeem_code(pool: &PgPool, code: &str, presented_client_id: &str) -> Response {
-    let (auth_state, producer_handle) = create_test_auth_state(pool.clone());
+    let (auth_state, producer_handle) = common::create_test_auth_state(pool.clone());
 
     let result = oauth::token(
-        test_tenant(),
+        common::test_tenant(),
         State(auth_state),
         oauth::TokenRequestBody(oauth::TokenRequest {
             grant_type: Some("authorization_code".to_string()),
@@ -149,7 +60,7 @@ async fn redeem_code(pool: &PgPool, code: &str, presented_client_id: &str) -> Re
 
     match result {
         Ok(response) => response,
-        Err(error) => error.into_response(),
+        Err(error) => axum::response::IntoResponse::into_response(error),
     }
 }
 
@@ -162,7 +73,7 @@ async fn response_json(response: Response) -> serde_json::Value {
 
 #[tokio::test]
 async fn redemption_by_a_different_client_is_rejected_and_leaves_the_code_unspent() {
-    let pool = setup_pool().await;
+    let pool = common::setup_oauth_test_db().await;
     let user_pubkey = Keys::generate().public_key().to_hex();
     let code = format!("binding-mismatch-{}", Uuid::new_v4());
     let email = format!("binding-mismatch-{}@example.com", Uuid::new_v4());
@@ -209,7 +120,7 @@ async fn redemption_by_a_different_client_is_rejected_and_leaves_the_code_unspen
 
 #[tokio::test]
 async fn redemption_by_the_issued_client_succeeds() {
-    let pool = setup_pool().await;
+    let pool = common::setup_oauth_test_db().await;
     let user_pubkey = Keys::generate().public_key().to_hex();
     let code = format!("binding-match-{}", Uuid::new_v4());
     let email = format!("binding-match-{}@example.com", Uuid::new_v4());
